# docs/master_swagger.yaml
basePath: /api/
definitions:
...
  curd.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  curd.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  curd.ReplyData-model_Gateway:
    properties:
      data:
        $ref: '#/definitions/types.Gateway'
      error:
        type: string
    type: object
  curd.ReplyData-model_History:
    properties:
      data:
        $ref: '#/definitions/types.History'
      error:
        type: string
    type: object
  curd.ReplyData-model_Notification:
    properties:
      data:
        $ref: '#/definitions/types.Notification'
      error:
        type: string
    type: object
...
  curd.ReplyList-model_Gateway:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Gateway'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_History:
    properties:
      data:
        items:
          $ref: '#/definitions/types.History'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  curd.ReplyList-model_Notification:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Notification'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.Gateway:
    properties:
      created:
        type: string
      desc:
        type: string
      disabled:
        type: boolean
      id:
        type: string
      name:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  types.History:
    properties:
      device:
        type: string
      device_id:
        type: string
      id:
        type: integer
      point:
        description: 数据点
        type: string
      time:
        type: string
      value:
        description: 值
        type: number
    type: object
...
  types.Notification:
    properties:
      alarm_id:
        type: string
      channels:
        items:
          type: string
        type: array
      created:
        type: string
      id:
        type: integer
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API文档
  title: 物联大师接口文档
  version: 3.2 版本
paths:
...
  /gateway/{id}:
    get:
      consumes:
      - application/json
      description: 获取网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 获取网关
      tags:
      - gateway
    post:
      consumes:
      - application/json
      description: 修改网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      - description: 网关信息
        in: body
        name: gateway
        required: true
        schema:
          $ref: '#/definitions/types.Gateway'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 修改网关
      tags:
      - gateway
  /gateway/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 删除网关
      tags:
      - gateway
  /gateway/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 禁用网关
      tags:
      - gateway
  /gateway/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用网关
      parameters:
      - description: 网关ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 启用网关
      tags:
      - gateway
  /gateway/count:
    post:
      consumes:
      - application/json
      description: 查询网关数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询网关数量
      tags:
      - gateway
  /gateway/create:
    post:
      consumes:
      - application/json
      description: 创建网关
      parameters:
      - description: 网关信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Gateway'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Gateway'
      summary: 创建网关
      tags:
      - gateway
...
  /gateway/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入网关
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 导入网关
      tags:
      - gateway
  /gateway/list:
    get:
      consumes:
      - application/json
      description: 查询网关
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Gateway'
      summary: 查询网关
      tags:
      - gateway
  /gateway/search:
    post:
      consumes:
      - application/json
      description: 查询网关
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Gateway'
      summary: 查询网关
      tags:
      - gateway
  /history/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除历史
      parameters:
      - description: 历史ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_History'
      summary: 删除历史
      tags:
      - history
  /history/count:
    post:
      consumes:
      - application/json
      description: 查询历史
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询历史数量
      tags:
      - history
...
  /history/list:
    get:
      consumes:
      - application/json
      description: 查询历史
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_History'
      summary: 查询历史
      tags:
      - history
  /history/search:
    post:
      consumes:
      - application/json
      description: 查询历史
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_History'
      summary: 查询历史
      tags:
      - history
  /notification/{id}/delete:
    get:
      description: 删除通知
      parameters:
      - description: 通知ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-model_Notification'
      summary: 删除通知
      tags:
      - notification
  /notification/count:
    post:
      consumes:
      - application/json
      description: 查询通知
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyData-int64'
      summary: 查询通知
      tags:
      - notification
  /notification/list:
    get:
      description: 查询通知
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Notification'
      summary: 查询通知
      tags:
      - notification
  /notification/search:
    post:
      consumes:
      - application/json
      description: 查询通知
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/curd.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/curd.ReplyList-model_Notification'
      summary: 查询通知
      tags:
      - notification
...
swagger: "2.0"